refactor: changed return value of problem to usize

+++ src/problem.rs
@@ -1,5 +1,5 @@
 pub trait Problem {
-    fn solve_part_one(&self, input: &str) -> u64;
-    fn solve_part_two(&self, input: &str) -> u64;
+    fn solve_part_one(&self, input: &str) -> usize;
+    fn solve_part_two(&self, input: &str) -> usize;
     fn index(&self) -> usize;
 }
